% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMolecule.R
\name{getMolecule}
\alias{getMolecule}
\alias{getMass}
\alias{getIsotope}
\alias{getFormula}
\alias{getScore}
\alias{getValid}
\title{Calculate mass and isotope information for a molecule given as sum formula}
\usage{
getMolecule(formula, elements = NULL, z = 0, maxisotopes = 10)

getMass(molecule)

getIsotope(molecule, index)

getFormula(molecule)

getScore(molecule)

getValid(molecule)
}
\arguments{
\item{formula}{Sum formula.}

\item{elements}{list of allowed chemical elements, defaults to full periodic system of elements.}

\item{z}{charge z of molecule for exact mass calculation.}

\item{maxisotopes}{maximum number of isotopes shown in the resulting molecules}

\item{molecule}{An initialized molecule as returned by getMolecule() or the decomposeMass() and decomposeIsotope() functions.}

\item{index}{Return the n-th isotope mass/abundance pair of the molecule}
}
\value{
A list with the elements `formula` repeated sum formula, `mass` exact 
    monoisotopic mass of molecule, `score` probability, for given molecules a 
    dummy value which is always 1.0, `valid` result of neutrogen rule check, 
    `isotopes` a list of isotopes.
}
\description{
Parse the sum formula and calculate the theoretical exact mass 
    and the isotope distribution.
}
\details{
\code{getMolecule()} will parse the sum formula and calculate the 
    theoretical exact monoisotopic mass and the isotope distribution. For a 
    given element, return the different mass values.
}
\examples{
# Ethanol
getMolecule("C2H6O")

}
\references{
For a description of the underlying IMS see citation("Rdisop")
}
\author{
Steffen Neumann <sneumann@IPB-Halle.DE>
}
