% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{isotopes}
\alias{isotopes}
\title{isotope information.}
\format{
A data frame of 4 columns for 398 chemical isotopes.
\describe{
\item{element}{The element name.}
\item{isotope}{The isotope name.}
\item{mass}{The absolute mass of this isotope in Dalton.}
\item{abundance}{The absolute abundance of this isotope.}
}
}
\source{
Isotope patterns were obtained through wikipedia.org.
}
\usage{
isotopes
}
\description{
isotope information.
}
\keyword{datasets}
