% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mono_masses}
\alias{mono_masses}
\title{Monoisotopic mass information..}
\format{
A named vector containing for each chemical element the most abundant isotope mass.
}
\source{
Isotope patterns were obtained through wikipedia.org.
}
\usage{
mono_masses
}
\description{
Monoisotopic mass information..
}
\keyword{datasets}
