% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposeIsotopes.R
\name{decomposeIsotopes}
\alias{decomposeIsotopes}
\alias{decomposeMass}
\alias{isotopeScore}
\title{Mass Decomposition of Isotope Patterns}
\usage{
decomposeIsotopes(
  masses,
  intensities,
  ppm = 2,
  mzabs = 1e-04,
  elements = NULL,
  filter = NULL,
  z = 0,
  maxisotopes = 10,
  minElements = "C0",
  maxElements = "C999999"
)

decomposeMass(
  mass,
  ppm = 2,
  mzabs = 1e-04,
  elements = NULL,
  filter = NULL,
  z = 0,
  maxisotopes = 10,
  minElements = "C0",
  maxElements = "C999999"
)

isotopeScore(
  molecule,
  masses,
  intensities,
  elements = NULL,
  filter = NULL,
  z = 0
)
}
\arguments{
\item{masses}{A vector of masses (or m/z values) of an isotope cluster.}

\item{intensities}{Absolute or relative intensities of the \code{masses} peaks.}

\item{ppm}{Allowed deviation of hypotheses from given mass.}

\item{mzabs}{Absolute deviation in Dalton (mzabs and ppm will be added).}

\item{elements}{List of allowed chemical elements, defaults to CHNOPS.}

\item{filter}{NYI, will be a selection of DU, DBE and Nitrogen rules.}

\item{z}{Charge z of m/z peaks for calculation of real mass, keep z=0 for auto-detection.}

\item{maxisotopes}{Maximum number of isotopes shown in the resulting molecules.}

\item{minElements}{Molecular formulas, which contain lower and upper boundaries of allowed formula respectively.}

\item{maxElements}{Molecular formulas, which contain lower and upper boundaries of allowed formula respectively.}

\item{mass}{A single mass (or m/z value).}

\item{molecule}{An initialized molecule as returned by getMolecule() or the decomposeMass() and decomposeIsotope() functions.}
}
\value{
A list of molecules, which contain the sub-lists `formulas` potential 
    formulas, `exactmass` exact mass of each hypothesis (not monoisotopic), 
    `score` calculated score, `isotopes` a list of isotopes.
}
\description{
Calculate the elementary compositions from an exact Mass or
    Isotope Pattern, obtained e.g. by FTICR or TOF mass spectrometers.
}
\details{
Sum formulas are calculated which explain the given mass or isotope pattern.
}
\examples{
# Glutamate: 
decomposeIsotopes(c(147.0529, 148.0563), c(100.0, 5.56))

}
\references{
For a description of the underlying IMS see citation("Rdisop")
}
\author{
Steffen Neumann <sneumann@IPB-Halle.DE>
}
